# src/main/java/com/example/bankcards/dto/UserResponse.java
package com.example.bankcards.dto;

import lombok.Data;
import java.util.UUID;

@Data
public class UserResponse {
    private UUID id;
    private String username;
    private String role;
    private String status;
}